orchestrator/requirements.txt
langgraph
langchain
langchain-anthropic
langchain-community
mcp
pydantic
fastapi
uvicorn
websockets
aiofiles
supabase
python-dotenv
langgraph-checkpoint
tavily-python